% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_from_lm.R
\name{bootstrap_from_lm}
\alias{bootstrap_from_lm}
\title{Bootstrap Confidence Intervals for Regression Coefficients}
\usage{
bootstrap_from_lm(input_lm, alpha = 0.05, B = 1000)
}
\arguments{
\item{input_lm}{A lm in which your model is fit.}

\item{alpha}{the level of the confidence interval. Default is 0.05}

\item{B}{The number of bootstrapped samples to conduct. Default is 1000.}
}
\value{
A table with coefficient estimates and confidence intervals
}
\description{
Bootstrap Confidence Intervals for Regression Coefficients
}
\examples{

require(BootPermLm)
require(stats)

x1 <- rnorm(100)
x2 <- rnorm(100) + x1
x3 <- rnorm(100)
y <- rnorm(100) + x1 + 2*x2
my_lm <- lm(y ~ x1 + x2 + x3)

bootstrap_from_lm(my_lm)



}
\references{
https://en.wikipedia.org/wiki/Bootstrapping_(statistics)
}
