% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_from_matrix.R
\name{bootstrap_from_matrix}
\alias{bootstrap_from_matrix}
\title{Bootstrap Confidence Intervals for Regression Coefficients}
\usage{
bootstrap_from_matrix(y, x, alpha = 0.05, B = 1000)
}
\arguments{
\item{y}{The response variables (as a matrix)}

\item{x}{The predictor variables (as a matrix). Note that if an intercept term is desired, you should include a column of 1s.}

\item{alpha}{the level of the confidence interval. Default is 0.05}

\item{B}{The number of permutations to conduct. Default is 1000.}
}
\value{
A table with coefficient estimates and confidence intervals
}
\description{
Bootstrap Confidence Intervals for Regression Coefficients
}
\examples{

require(BootPermLm)
require(stats)
x1 <- rnorm(100)
x2 <- rnorm(100) + x1
x3 <- rnorm(100)
y <- rnorm(100) + x1 + 2*x2
x <- cbind("(Intercept)" = 1, x1, x2, x3)


bootstrap_from_matrix(y, x)



}
\references{
https://en.wikipedia.org/wiki/Bootstrapping_(statistics)
}
