% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_test_lm.R
\name{bootstrap_test_lm}
\alias{bootstrap_test_lm}
\title{Bootstrap Standard Errors for Regression Coefficients}
\usage{
bootstrap_test_lm(y, x, alpha = 0.05, B = 1000)
}
\arguments{
\item{y}{The response variables}

\item{x}{The predictor variables}

\item{alpha}{the level of the confidence interval. Default is 0.05}

\item{B}{The number of permutations to conduct. Default is 1000.}
}
\value{
A table with coefficient estimates and confidence intervals
}
\description{
Bootstrap Standard Errors for Regression Coefficients
}
\examples{

require(BootPermLm)
require(stats)
x1 <- rnorm(100)
x2 <- rnorm(100) + x1
x3 <- rnorm(100)
y <- rnorm(100) + x1 + 2*x2
x <- cbind(x1, x2, x3)


bootstrap_test_lm(y, x)



}
\references{
https://en.wikipedia.org/wiki/Bootstrapping_(statistics)
}
