% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_matrix.R
\name{permutation_test_matrix}
\alias{permutation_test_matrix}
\title{Permutation Test Regression Coefficients}
\usage{
permutation_test_matrix(y, x, B = 1000)
}
\arguments{
\item{y}{The response variables (As a matrix)}

\item{x}{The predictor variables  (As a matrix). Note that if an intercept term is desired, this matrix should contain a column of 1s}

\item{B}{The number of permutations to conduct. Default is 1000.}
}
\value{
A table with coefficient estimates and permutation p-values.
}
\description{
Permutation Test Regression Coefficients
}
\examples{

require(BootPermLm)
require(stats)
x1 <- rnorm(100)
x2 <- rnorm(100) + x1
x3 <- rnorm(100)
y <- rnorm(100) + x1 + 2*x2
x <- cbind("(Intercept)" = 1, x1, x2, x3)


permutation_test_matrix(y, x)



}
\references{
https://en.wikipedia.org/wiki/Permutation_test
}
