% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_lm.R
\name{permutation_test_lm}
\alias{permutation_test_lm}
\title{Permutation Test Regression Coefficients}
\usage{
permutation_test_lm(input_lm, B = 1000)
}
\arguments{
\item{input_lm}{A lm in which your model is fit.}

\item{B}{The number of permutations to conduct. Default is 1000.}
}
\value{
A matrix with coefficient estimates and permutation p-values.
}
\description{
Permutation Test Regression Coefficients
}
\examples{

require(BootPermLm)
require(stats)
x1 <- rnorm(100)
x2 <- rnorm(100) + x1
x3 <- rnorm(100)
y <- rnorm(100) + x1 + 2*x2
my_lm <- lm(y ~ x1 + x2 + x3)

permutation_test_lm(my_lm)





}
\references{
https://en.wikipedia.org/wiki/Permutation_test
}
